MongoDB integration OK.
