+++ coffee-web/pom.xml
@@ -26,8 +26,10 @@
 	</properties>
 
 	<dependencies>
-		<!-- <dependency> <groupId>org.springframework.boot</groupId> <artifactId>spring-boot-starter-data-mongodb</artifactId> 
-			</dependency> -->
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-data-mongodb</artifactId>
+		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
@@ -35,12 +37,6 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.jongo</groupId>
-			<artifactId>jongo</artifactId>
-			<version>1.3.0</version>
 		</dependency>
 
 		<dependency>
@@ -56,7 +52,7 @@
 	</dependencies>
 
 	<build>
-	 	<finalName>${project.artifactId}</finalName>
+		<finalName>${project.artifactId}</finalName>
 		<plugins>
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
@@ -69,10 +65,6 @@
 				<version>0.4.13</version>
 				<configuration>
 					<imageName>${project.artifactId}</imageName>
-					<!-- 
-					<baseImage>openjdk:8-jre</baseImage>
-					<entryPoint>["java", "-jar", "/${project.build.finalName}.jar"]</entryPoint>
-					 -->
 					<dockerDirectory>docker</dockerDirectory>
 					<resources>
 						<resource>
