--- coffee-web/src/main/java/me/dkist/iot/web/service/PersonService.java
@@ -1,17 +1,0 @@
-package me.dkist.iot.web.service;
-
-import org.jongo.MongoCollection;
-import org.springframework.stereotype.Service;
-
-@Service
-public class PersonService {
-
-	//private final static Jongo jongo = new Jongo(new MongoClient().getDB("iot-coffee"));
-	
-	public final MongoCollection getCollection() {
-		//return jongo.getCollection("person");
-		return null;
-	}
-	
-	
-}
